+++ contracts/RMRK/utils/RMRKEquipRenderUtils.sol
@@ -149,7 +149,7 @@ contract RMRKEquipRenderUtils is
         uint64[] memory assets = target_.getActiveAssets(tokenId);
         uint16[] memory priorities = target_.getActiveAssetPriorities(tokenId);
         uint256 len = assets.length;
-        if (len == 0) {
+        if (len == uint256(0)) {
             revert RMRKTokenHasNoAssets();
         }
 
@@ -209,7 +209,7 @@ contract RMRKEquipRenderUtils is
 
         uint64[] memory assets = target_.getPendingAssets(tokenId);
         uint256 len = assets.length;
-        if (len == 0) {
+        if (len == uint256(0)) {
             revert RMRKTokenHasNoAssets();
         }
